Init.d Script: Do not handle TMP folder mounts after boot The reason why we want to handle the tmp folders AFTER boot is so that the package manager get's a chance to create them after a factory reset.
But as of Android 4.2, we can no longer handle mounts from within the app as these mounts will only affect the app that creates them. There has been gound solutions to this issue, which might be added to some custom ROM's, but we want these mounts to work on all ROM's with root and init.d

So we move these mounts into the boot part of the script. This is executed from within init.rc and will add the mounts on a global scale. This just means that people will have to do one reboot after a factory reset in order to get /data/data/com.android.providers.downloads/cache handled corectly