Make better compatibility between different devices and ROM's One of the biggest issues with different devices and ROM's, is that they are using different shells with different support. Some ROM's uses a pure sh, others link sh to busybox, some to mkfs, some to buxybox mkfs and even a few uses dash.
These different environments is the cause of the majority of issues, although the script tries to adapt to the shell it is running in. However,
the script does still need some features, and when running a shell with almost none at all (Dash for one), it is dificult to add support for it.
To avoid this in the future, we now add a small runner script to init.d and move Mounts2SD out of the init.d folder. The runner script will invoke the upstart script
using the best available shell on the device, and limiting these issues by using simular shells on all devices. This will create a more controlled environment.

We also revert back to our prev busybox binary, as this has proven to cause less issues.